+++ projects/tokens/contracts/erc-1155/erc-1155-meta/ERC1155RaribleMeta.sol
@@ -32,6 +32,18 @@ contract ERC1155RaribleMeta is ERC1155Base, IsPrivateCollection, MinterAccessCon
         emit CreateERC1155Rarible(_msgSender(), _name, _symbol);
     }
 
+    function __ERC1155Rarible_init_proxy(string memory _name, string memory _symbol, string memory baseURI, string memory contractURI, address transferProxy, address lazyTransferProxy, address initialOwner) external {
+        __ERC1155Rarible_init_unchained(_name, _symbol, baseURI, contractURI, transferProxy, lazyTransferProxy);
+
+        __MetaTransaction_init_unchained("ERC1155RaribleMeta", "1");
+
+        isPrivate = false;
+
+        transferOwnership(initialOwner);
+
+        emit CreateERC1155Rarible(_msgSender(), _name, _symbol);
+    }
+
     function _msgSender() internal view virtual override(ContextUpgradeable, EIP712MetaTransaction) returns (address payable) {
         return super._msgSender();
     }