+++ projects/tokens/test-mint-721.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 truffle test ./test/erc-721-minimal/Mint721Validator.test.js \
             ./test/contracts/erc-721/Mint721ValidatorTest.sol \
-            ./test/contracts/TestERC1271.sol 
+            ./test/contracts/TestERC1271.sol --compile-all