+++ projects/deploy/CHANGELOG.md
@@ -43,6 +43,26 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 - remove auction: not needed at the moment ([ecbc515](https://github.com/rariblecom/protocol-contracts/commit/ecbc515b058acadeace296ede9793afe5831bfcf))
 - removed uneeded code ([b5110cc](https://github.com/rariblecom/protocol-contracts/commit/b5110cc7a0b712ce16339756e6e3f190e4a8d443))
 
+**Note:** Version bump only for package @rarible/deploy
+
+## [0.8.6](https://github.com/rariblecom/protocol-contracts/compare/v0.8.1...v0.8.6) (2024-04-27)
+
+### Bug Fixes
+
+- build ([5a04947](https://github.com/rariblecom/protocol-contracts/commit/5a049478b8acde4a8681b09d2f9aa5c54bd9fa58))
+
+### Features
+
+- add build deps ([822c123](https://github.com/rariblecom/protocol-contracts/commit/822c123199ac04598b6f8ab995ee8e8a66828015))
+- PT-4662 get rid of nothing to fill tests ([4059dcb](https://github.com/rariblecom/protocol-contracts/commit/4059dcb20c100e159643e40929392b05f90f7038))
+- PT-5072 add seaport 1.6 to exchange wrapper, deploy wrapper to sepolia ([c6f10f1](https://github.com/rariblecom/protocol-contracts/commit/c6f10f151740dd76da0e997b8bf499f03cba00a9))
+
+## [0.8.5](https://github.com/rariblecom/protocol-contracts/compare/v0.8.1...v0.8.5) (2024-04-27)
+
+### Bug Fixes
+
+- add missing import ([ec5a4ea](https://github.com/rariblecom/protocol-contracts/commit/ec5a4ea5a59c98a65aad20a2b382d8eb8ec50771))
+
 ### Features
 
 - add build deps ([822c123](https://github.com/rariblecom/protocol-contracts/commit/822c123199ac04598b6f8ab995ee8e8a66828015))